templates/context-engine/requirements-context.txt
llama-index-core>=0.11.0
llama-index-vector-stores-chroma>=0.2.0
llama-index-embeddings-huggingface>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pyyaml>=6.0
networkx>=3.0
tiktoken>=0.7.0
mcp>=1.20.0
